fix error for supported nfs versions
[#186911574]

Our docs say that `3.0` is a valid nfs version. It turns out that this was
undefined behavior in old versions of nfs-utils and will not work when
the underlying CF Deployments runs on Jammy (as opposed to Xenial).

Users have pointed reported:
after updating the CF deployment from a xenial to a jammy stemcell, previously
working volume services ( created with `version: 3.0`) stopped working.
After updating their services to use `version: 3` keeping the rest of
the configuration identical, the mounts succeeded again.

The best proof for this change (without trying to track a commit in nfs-utils source),
I could find is a comment by Steve Dickson (the maintainer) here:
https://bugzilla.redhat.com/show_bug.cgi?id=1522633

```
There are no minor versions with NFS v3
```